feat(common): update register page
