+++ src/ui/components/register/Register.css
@@ -1,28 +1,28 @@
 body {
-    background-color: rgb(76, 76, 76);
+    background-color: rgb(83, 83, 83);
 }
 
 .form-register {
-    width: 558px;
-    height: 670px;
+    width: 34.875rem;
+    height: 41.875rem;
     margin: 2% auto;
     box-sizing: border-box;
     position: absolute;
-    left: 0px;
-    right: 0px;
-    top: 20px;
-    bottom: 0px;
+    left: 0rem;
+    right: 0rem;
+    top: 1.25rem;
+    bottom: 0rem;
     background: #FFFFFF;
-    border: 1px solid #DFE0EB;
-    border-radius: 8px;
+    border: 0.063rem solid #DFE0EB;
+    border-radius: 0.5rem;
 }
 
 .sumbit-register {
     display: inline-block;
     color: #f6f6f7;
-    width: 525px;
-    height: 55px;
-    padding: 0 20px;
+    width: 32.813rem;
+    height: 3.438rem;
+    padding: 0 1.25rem;
     background: rgb(78, 119, 252);
     border-radius: 5px;
     outline: none;
@@ -31,16 +31,16 @@ body {
     text-align: center;
     transition: all 0.2s linear;
     margin: 7% auto;
-    letter-spacing: 0.05em;
-    margin-top: 5px;
-    margin-left: 15px;
-    margin-bottom: 0px;
+    letter-spacing: 0.05rem;
+    margin-top: 0.313rem;
+    margin-left: 0.938rem;
+    margin-bottom: 0rem;
 }
 
 .title-one {
     text-align: center;
     color: #8a8b94;
-    margin-top: 0px;
+    margin-top: 0rem;
     font-size: 20px;
     
 }
@@ -49,8 +49,8 @@ body {
     font-family: 'Playfair Display', serif;
     text-align: center;
     color: #8a8b94;
-    margin-top: 10px;
-    margin-bottom: 50px;
+    margin-top: 0.625rem;
+    margin-bottom: 3.125rem;
     
 }
 
@@ -62,35 +62,35 @@ body {
 
 .logo {
     font-family: 'Playfair Display', serif;
-    width: 50px;
-    height: 50px;
+    width: 3.125rem;
+    height: 3.125rem;
     display: block;
-    margin: 0px auto;
-    margin-top: 10px;
-    margin-bottom: 0px;
+    margin: 0rem auto;
+    margin-top: 0.625rem;
+    margin-bottom: 0rem;
 }
 
 .label-input-error{
     text-transform: uppercase;
     text-align: center;
     display: block;
-    margin: 0 0 10px;
+    margin: 0 0 0.625rem;
     color: rgb(215, 8, 8);
     font-size: 12px;
     font-weight: semibold;
     line-height: 1;
     letter-spacing: .2em;
-    margin-top: -30px;
-    margin-bottom: 20px;
-    margin-left: 20px;
+    margin-top: -1.875;
+    margin-bottom: 1.25rem;
+    margin-left: 1.25rem;
 }
 
 .logo2 {
     font-family: 'Playfair Display', serif;
-    width: 200px;
-    height: 200px;
+    width: 12.5rem;
+    height: 12.5rem;
     display: block;
-    margin: 0px auto;
-    margin-top: 10px;
-    margin-bottom: 0px;
+    margin: 0rem auto;
+    margin-top: 0.625rem;
+    margin-bottom: 0rem;
 }
