+++ src/ui/components/register/subcomponents/Title/Title.css
@@ -4,5 +4,5 @@
     font-weight: bold;
     color: #3e403f;
     text-align: center;
-    margin-bottom: 0px;
+    margin-bottom: 0rem;
 }
